# docs/reference/openapi.yaml
openapi: 3.0.3
info:
  title: DragonCP API
  description: |
    Machine-readable index of the DragonCP backend HTTP API.

    **`api.md` in this directory is authoritative.** This file exists so the API
    can be imported into an API client or used to generate a stub client. It
    lists every endpoint with its method, path, auth requirement and a one-line
    summary; request and response schemas are deliberately thin and name only
    the common fields. Where the two differ, `api.md` is correct and this file
    needs fixing.

    Written by reading the `@app.route` / `@*_bp.route` decorators in `app.py`
    and `routes/`. Every blueprint is registered under the `/api` prefix, which
    is part of the server URL below, so paths here are written without it.

    `GET /` is not listed: it renders the legacy HTML page and is not an API
    route.

    Endpoint count: 95 method+path operations.

    ## Authentication

    Most endpoints require a JWT access token issued by `POST /auth/login`:

    ```
    Authorization: Bearer <access-token>
    ```

    Query-string tokens (`?token=...`) are not accepted on normal HTTP requests.

    The three webhook receivers do not take a JWT. They are guarded by
    `require_webhook_auth`, which enforces an HMAC `X-DragonCP-Signature` header
    when `WEBHOOK_SECRET` is set and a source-IP allowlist when
    `WEBHOOK_ALLOWED_IPS` is set. When both are set, either check passing is
...
    warning.
  version: "2.0.0"
  contact:
...
servers:
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from POST /auth/login.
    webhookSignature:
      type: apiKey
      in: header
      name: X-DragonCP-Signature
      description: |
        HMAC-SHA256 of the raw request body, formatted `sha256=<hex digest>`,
        keyed with WEBHOOK_SECRET. Enforced only when WEBHOOK_SECRET is set. A
        request from an address in WEBHOOK_ALLOWED_IPS is accepted without it.
...
  schemas:
    SuccessResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success]
        message:
          type: string
      required: [status]
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          enum: [error]
        message:
          type: string
      required: [status]
    ConfigObject:
      type: object
      description: |
        Raw configuration key/value object, not wrapped in a status envelope.
        Secrets are stripped before the response is sent.
      additionalProperties: true
...
  responses:
    Success:
      description: Success
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SuccessResponse'
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid access token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
paths:
...
  /config:
    get:
      tags: [Configuration]
      summary: Return the active runtime configuration (environment plus session overrides)
      responses:
        '200':
          description: Configuration object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigObject'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Configuration]
      summary: Apply session-level configuration overrides
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigObject'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /config/reset:
    post:
      tags: [Configuration]
      summary: Drop session overrides and fall back to environment values
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /config/env-only:
    get:
      tags: [Configuration]
      summary: Return environment configuration only, ignoring session overrides
      responses:
        '200':
          description: Configuration object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigObject'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /connect:
    post:
      tags: [Configuration]
      summary: Open an SSH connection to the remote server and attach it to the runtime
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                host:
                  type: string
                username:
                  type: string
                password:
                  type: string
                key_path:
                  type: string
      responses:
        '200':
          description: |
            Connection attempted. A missing field or a failed connection is
            reported in the body as `status: error`, still with HTTP 200.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Body was valid JSON but not an object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /disconnect:
    post:
      tags: [Configuration]
      summary: Close the current SSH connection and clear connection state
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auto-connect:
    get:
      tags: [Configuration]
      summary: Connect over SSH using the configured environment credentials
      responses:
        '200':
          description: |
            Connection attempted. Missing credentials or a failed connection are
            reported in the body as `status: error`, still with HTTP 200.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /ssh-config:
    get:
      tags: [Configuration]
      summary: Return the loaded SSH settings, with the password reduced to a boolean
      responses:
        '200':
          description: SSH settings
          content:
            application/json:
              schema:
                type: object
                properties:
                  host:
                    type: string
                  username:
                    type: string
                  key_path:
                    type: string
                  has_password:
                    type: boolean
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ==================== MEDIA BROWSING ====================
...
  /logs:
    get:
      tags: [Logs]
      summary: Return recent backend log records matching the requested severity
      description: >-
        Matching is not uniformly "level and above". ALL returns everything,
        ERROR returns ERROR and CRITICAL, WARNING returns WARNING, ERROR and
        CRITICAL, and DEBUG, INFO and CRITICAL match that level exactly. An
        unrecognised value is coerced to ERROR.
      parameters:
        - name: level
          in: query
          required: false
          schema:
            type: string
            enum: [DEBUG, INFO, WARNING, ERROR, CRITICAL, ALL]
            default: ERROR
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 200
            maximum: 1000
        - name: search
          in: query
          required: false
          description: Case-insensitive substring match.
          schema:
            type: string
      responses:
        '200':
          description: Log records
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  log_file:
                    type: string
                  level:
                    type: string
                  limit:
                    type: integer
                  line_count:
                    type: integer
                  size_bytes:
                    type: integer
                  last_modified:
                    type: string
                    format: date-time
                  lines:
                    type: array
                    description: Oldest first.
                    items:
                      type: object
                      properties:
                        level:
                          type: string
                        text:
                          type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
...